add get form by id

+++ src/main/java/com/blackstone/webappsorganizationsurvey/repository/FormRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface FormRepository extends JpaRepository<Form, Integer> {
+public interface FormRepository extends JpaRepository<Form, Long> {
 }
